+++ dev-site-config/generatedComponentConfig.js
@@ -103,6 +103,7 @@ import BackCloseActionHeader from '../packages/terra-action-header/examples/test
 import BackClosePreviousNextActionHeader from '../packages/terra-action-header/examples/test-examples/BackClosePreviousNextActionHeader.example';
 import CloseActionHeader from '../packages/terra-action-header/examples/test-examples/CloseActionHeader.example';
 import DefaultActionHeader from '../packages/terra-action-header/examples/test-examples/DefaultActionHeader.example';
+import EndElementActionHeader from '../packages/terra-action-header/examples/test-examples/EndElementActionHeader.example';
 import MaximizeCloseActionHeader from '../packages/terra-action-header/examples/test-examples/MaximizeCloseActionHeader.example';
 import MinimizeCustomButtonActionHeader from '../packages/terra-action-header/examples/test-examples/MinimizeCustomButtonActionHeader.example';
 import WrappedTitleActionHeader from '../packages/terra-action-header/examples/test-examples/WrappedTitleActionHeader.example';
@@ -372,10 +373,10 @@ import LightOverlay from '../packages/terra-overlay/examples/test-examples/Overl
 import OnRequestCloseOverlay from '../packages/terra-overlay/examples/test-examples/Overlay/OnRequestCloseOverlay.example';
 import DefaultOverlayContainer from '../packages/terra-overlay/examples/test-examples/OverlayContainer/DefaultOverlayContainer.example';
 import ProfileImageDefault from '../packages/terra-profile-image/tests/nightwatch/ProfileImageDefault';
-import ProgressBarGradient from '../packages/terra-progress-bar/tests/nightwatch/ProgressBarGradient';
-import ProgressBarColor from '../packages/terra-progress-bar/tests/nightwatch/components/ProgressBarColor';
-import ProgressBarDefault from '../packages/terra-progress-bar/tests/nightwatch/components/ProgressBarDefault';
-import ProgressBarSize from '../packages/terra-progress-bar/tests/nightwatch/components/ProgressBarSize';
+import ProgressBarColor from '../packages/terra-progress-bar/examples/test-examples/ProgressBarColor.example';
+import ProgressBarDefault from '../packages/terra-progress-bar/examples/test-examples/ProgressBarDefault.example';
+import ProgressBarGradient from '../packages/terra-progress-bar/examples/test-examples/ProgressBarGradient.example';
+import ProgressBarSize from '../packages/terra-progress-bar/examples/test-examples/ProgressBarSize.example';
 import MockPropsTable from '../packages/terra-props-table/tests/nightwatch/MockPropsTable';
 import MockPropsTableWithTitle from '../packages/terra-props-table/tests/nightwatch/MockPropsTableWithTitle';
 import DefaultResponsiveElement from '../packages/terra-responsive-element/tests/nightwatch/DefaultResponsiveElement';
@@ -675,6 +676,11 @@ const componentConfig = {
         name: 'Default Action Header',
         path: '/default-action-header',
         component: DefaultActionHeader,
+      },
+      {
+        name: 'End Element Action Header',
+        path: '/end-element-action-header',
+        component: EndElementActionHeader,
       },
       {
         name: 'Maximize Close Action Header',
@@ -2400,11 +2406,6 @@ const componentConfig = {
     path: '/progress-bar',
     tests: [
       {
-        name: 'Progress Bar Gradient',
-        path: '/progress-bar-gradient',
-        component: ProgressBarGradient,
-      },
-      {
         name: 'Progress Bar Color',
         path: '/progress-bar-color',
         component: ProgressBarColor,
@@ -2413,6 +2414,11 @@ const componentConfig = {
         name: 'Progress Bar Default',
         path: '/progress-bar-default',
         component: ProgressBarDefault,
+      },
+      {
+        name: 'Progress Bar Gradient',
+        path: '/progress-bar-gradient',
+        component: ProgressBarGradient,
       },
       {
         name: 'Progress Bar Size',